refactor(encoding): extract cost model to cost_model/ module
Move `HcOptState`, `HcOptimalCostProfile`, `HcOptPriceType`,
`HcDictEntropySeed`, and the per-symbol price constants from the
9000+ line `match_generator.rs` monolith into a dedicated `cost_model/`
submodule.

Mechanical move only — no behaviour change:

- All extracted items keep their original names and signatures.
- `pub(crate)` visibility on the moved items so `match_generator`
  callers can import them via `super::cost_model::*`.
- `HcParseMode` (still in `match_generator`) is now `pub(crate)` so the
  extracted `HcOptimalCostProfile::for_mode` can pattern-match on it.

`match_generator.rs` drops 521 lines (from 10613 → 10092) and gets the
import block at the top. Build, clippy (x86_64 + aarch64), and the
correctness gate (roundtrip + cross-validation + ratio gate) all stay
green.

Part of #111 Phase 1.
